its working!
appears to be working moderately well.  had to import a contentEditable plugin to make it work properly.  State is stored locally which is fine for a simple plugin, will need to be hooked up to redux somehow for larger applications like firestarter.  need to get remove this marker to work as well